Start working on ConfigDB dump format version 2
We need to be able to create (R1) classes via the `objects` key, which
v1 dumps do not allow. Eventually v2 dumps want to be able to resolve
loaded objects in any order; for now we are relying on the load order,
which means relying on key order preservation in JS and YAML.

Version 2 dumps will always be authoritative about the information they
contain. There will not be the option to choose not to overwrite
existing information. The /v1/load endpoint on the ConfigDB will
continue to accept the parameter for compatibility; the new /load
endpoint will not.
