+++ acs-configdb/lib/load.js
@@ -20,16 +20,6 @@ export class Loader {
 
     async load(req, res) {
         const dump = req.body;
-        const booleans = {
-            "true": true, "false": false,
-            "1": true, "0": false,
-            on: true, off: false,
-            yes: true, no: false,
-        };
-
-        const overwrite = booleans[req.query.overwrite];
-        if (overwrite == undefined)
-            return res.status(400).end();
 
         if (!await this.model.dump_validate(dump))
             return res.status(400).end();
@@ -50,7 +40,7 @@ export class Loader {
             }
         }
 
-        const st = await this.model.dump_load(dump, !!overwrite);
+        const st = await this.model.dump_load(dump);
         res.status(st).end();
     }
 }
