+++ acs-service-setup/dumps/sparkplug.yaml
@@ -1,13 +1,12 @@
 # ACS-specific deployment decisions
 service: !u UUIDs.Service.ConfigDB
-version: 1
-classes:
-  - !u UUIDs.Class.EdgeAgent
-  # Not created here
-  - !u ACS.Class.ClientRole
+version: 2
 objects:
   !u ConfigDB.Class.R2Class:
     - !u Auth.Class.PrincipalGroup
+  !u UUIDs.Class.Class:
+    - !u UUIDs.Class.EdgeAgent
+    - !u ACS.Class.ClientRole
   !u UUIDs.Class.App:
     - !u Edge.App.AgentConfig
   !u ACS.Class.ClientRole:
